+++ app/src/main/java/vn/poly/personalmanagement/ui/fragment/money/expenses/AddExpenseFragment.java
@@ -91,7 +91,7 @@ public class AddExpenseFragment extends Fragment
     private void addExpenses(){
         String title =edtTitle.getText().toString().trim();
         String sAmount = edtAmount.getText().toString().trim();
-        double amount =0;
+        long amount =0;
         String description = edtDescribe.getText().toString().trim();
         String date = CurrentDateTime.getCurrentDate();
         String time = CurrentDateTime.getCurrentTime();
@@ -104,7 +104,7 @@ public class AddExpenseFragment extends Fragment
             edtAmount.setError("Mời nhập số tiền đã chi");
             edtAmount.setFocusable(true);
             return;
-        }else amount= Double.parseDouble(sAmount);
+        }else amount= Long.parseLong(sAmount);
         if (description.isEmpty()){
             description=" ";
         }
